+++ packages/admin/src/components/entities/vols/common-edit/sections/index.ts
@@ -2,7 +2,7 @@ export { SidebarNavigation } from './sidebar-navigation';
 export { PreviousYearsSection } from './previous-years-section';
 export { AdditionalSection } from './additional-section';
 export { CustomFieldsSection } from './custom-fields-section';
-export { BadgeSection } from './badge-section';
+export { PersonalInfoSection } from './personal-info-section';
 export { ArrivalsSection } from './arrivals-section';
 export { HrInfoSection } from './hr-info-section';
-export { PersonalInfoSection } from './personal-info-section';
+export { VolInfoSection } from './vol-info-section';
